+++ CourseWorke/CourseWorke/Sailboat.h
@@ -9,8 +9,8 @@ private:
 public:
     Sailboat();
     Sailboat(std::string name, double maxSpeed, int passengerCapacity, int sailArea);
-    void SetSailArea(double sailArea);
-    double GetSailArea();
+    void SetSailArea(int sailArea);
+    int GetSailArea();
     void Print();
     void PrintToFile(std::ofstream& outputFile);
     void ReadFromFile(std::ifstream& inFile);
